+++ deploy/node/ppos_7.yml
@@ -1,0 +1,92 @@
+collusion:
+- host: 192.168.9.221
+  port: 16789
+  rpcport: 6789
+  sshport: 22
+  username: platon
+  password: Platon123!
+  id: 0b08d050d1e87c9a08b1d7236d8301d5bfa841a1666cd2f1a1e1d668683b755015e6734a66776e1934b4d08b2a72d961e86849ae197af34a705d73615f5a1dcb
+  nodekey: 747988b2a93122d777ab858bb9ddf6ef4de37be7bc0aff5da3f2d8b668db49b9
+  url: http://192.168.9.221:6789
+- host: 192.168.9.222
+  port: 16789
+  rpcport: 6789
+  sshport: 22
+  username: platon
+  password: Platon123!
+  id: 8c33688b6470bf3a79fcb93ede98370145d2f8802d6e1d3ae8598cb04c8627a11a92c8fc984282e16d6be6ac3e7feeb50bb2c60c960bc12e8e8eceea0b1edb79
+  nodekey: 66f039298545da5db387e94ce9798cac549e4b5ecae5ce6f57c63e86808b2586
+  url: http://192.168.9.222:6789
+- host: 192.168.9.223
+  port: 16789
+  rpcport: 6789
+  sshport: 22
+  username: platon
+  password: Platon123!
+  id: cb1c4b3658d9de93dfba2b7f25078e5ec0c101a6613ee11bfa56bc94583f9fe16261a2d53f245b3b1f374d133bdfe31f5421aea8e26dddc73bb3a29f5c91082c
+  nodekey: a98b7e6ca795554ab15f2aee48fc5f7d1cc38014ac6cf0f979c1e3b496852bdf
+  url: http://192.168.9.223:6789
+- host: 192.168.9.224
+  port: 16789
+  rpcport: 6789
+  sshport: 22
+  username: platon
+  password: Platon123!
+  id: cecc907b89397aad3ced544ac7e9237f9095259bb0850a9304046583b6cd040c40e3dd0427c1087bd543fa6bc3ee9766d216d59810457be2db3227cc9b0e9723
+  nodekey: 03af4b161f1ccac1bb5744f243df0411a47b150134bb2334ac6d3bf407f58442
+  url: http://192.168.9.224:6789
+nocollusion:
+- host: 192.168.9.225
+  port: 16789
+  rpcport: 6789
+  sshport: 22
+  username: platon
+  password: Platon123!
+  id: 458d4ee4a5212d666fe995e97dbe7e95b6de721b049d8dac3609dc376bc6badd92569a23855bbeee8be10ef1c594288e5f9df8e8a9bed042f9f2ff349bb9e102
+  nodekey: 6831b33c051cbe2bf3e0ec151bef085cf4c0f26240c40e60ee9bc82724c569c8
+  url: http://192.168.9.225:6789
+- host: 192.168.9.226
+  port: 16789
+  rpcport: 6789
+  sshport: 22
+  username: platon
+  password: Platon123!
+  id: 4bc4d3040a72dea70b3cf57efae73ae3fb9e2ec7513d1e3607ae2d10a7b0df3c8fcfc65bb8ed9fdd612d52da74c5f0ae187114f8e925273270c5c23c354cab59
+  nodekey: 45ee809930ef280190d03143eb934733428a0818c9dc3a1f5bf34c9a60854c92
+  url: http://192.168.9.226:6789
+- host: 192.168.9.227
+  port: 16789
+  rpcport: 6789
+  sshport: 22
+  username: platon
+  password: Platon123!
+  id: 963063a1054f8b4873e281d4b7868679702c093c4a0fe5f47f7e4abf7c372d0854ab54e3998fec7ea5c741020697fb34c0f7f46d844eef3c3f81d1f4f6b5d7eb
+  nodekey: 4555f773b32bc949faffcdfda623e326222ce0410c80ae39437ee91c52abe00f
+  url: http://192.168.9.227:6789
+- host: 192.168.9.225
+  port: 16788
+  rpcport: 6788
+  sshport: 22
+  username: platon
+  password: Platon123!
+  url: http://192.168.9.225:6788
+  id: bdda8529dcebf6071d0682647599007d626b08ff22a30bd8afd582f6325e2686dcc6bf6e1fd37f5fd85dd6c03365f3319f5f5fe217804ad6b9d728bc324a735a
+  nodekey: aeee432f300c7bce05670a3b38d2db39a160ca9eefdf32909d028e5dfdaeca06
+- host: 192.168.9.226
+  port: 16788
+  rpcport: 6788
+  sshport: 22
+  username: platon
+  password: Platon123!
+  url: http://192.168.9.226:6788
+  id: fdb37ad70acd5107a8386c4ba080551fb87c7e85e040f8f20bd03fd738ce7fd340f6ce0a3352acb7ebe0480b284ece8d3a586bf54bba5d0033573dbc1b7d631a
+  nodekey: c7a780a1c21a9b1cf1d9351cfed05a62d8e4a0d9b6b3b4d4717de875f430fda5
+- host: 192.168.9.227
+  port: 16788
+  rpcport: 6788
+  sshport: 22
+  username: platon
+  password: Platon123!
+  url: http://192.168.9.227:6788
+  id: f16a4a0f86e2830305963cd9a18b65557c9f518633395e5c926090141847f488ad4ea0012234a4d9b7f1e8f25fdbb23286cbd151d430a27d3619db18edfc61f9
+  nodekey: 7e167f15cb07beaa7c6f570f9647d92c1a753abaaae6552f93adb6fee9da369b
