+++ deploy/rely/config.json
@@ -1,0 +1,191 @@
+{
+    "eth":{
+		"SyncMode": "full",
+		"CbftConfig":{
+			"Period":           	1,
+			"Epoch":            	250000,
+			"MaxLatency":       	600,
+			"LegalCoefficient": 	1.0,
+			"Duration":         	10,
+			"BlockInterval":    	100,
+			"BreakpointType":		"",
+			"PeerMsgQueueSize": 	1024,
+			"EvidenceDir": 			"evidenceDir",
+			"MaxResetCacheSize": 	512,
+			"MaxQueuesLimit":		4096,
+			"MaxBlockDist":			192,
+			"MaxPingLatency":		5000,
+			"MaxAvgLatency":		2000,
+			"CbftVersion":			1,
+			"Remaining":			50000
+		},
+		"NetworkId":     	 1,
+		"LightServ": 	 	 0,
+		"LightPeers":    	 100,
+		"SkipBcVersionCheck":false,
+		"DatabaseHandles":	 0,
+		"DatabaseCache": 	 768,
+		"NoPruning":	     false,
+		"TrieCache":     	 256,
+		"TrieTimeout":   	 3600000000000,
+		"BodyCacheLimit": 	 256,
+		"BlockCacheLimit":	 256,
+		"MaxFutureBlocks":	 256,
+		"BadBlockLimit":  	 10,
+		"TriesInMemory":  	 128,
+		"BlockChainVersion": 3,
+		"SoftResponseLimit": 2097152,
+		"EstHeaderRlpSize":  500,
+		"DefaultTxsCacheSize":20,
+		"DefaultBroadcastInterval":100000000,
+		"Etherbase":	 	 "0x0000000000000000000000000000000000000000",
+		"MinerGasFloor": 	 3150000000,
+		"MinerGasCeil":  	 3150000000,
+		"MinerGasPrice": 	 100000000,
+		"MinerRecommit": 	 3000000000,
+		"MinerNoverify": 	 false,
+		"MiningLogAtDepth": 		7,
+		"TxChanSize":				4096,
+		"ChainHeadChanSize":		10,
+		"ChainSideChanSize":		10,
+		"ResultQueueSize":		 	10,
+		"ResubmitAdjustChanSize":	10,
+		"MinRecommitInterval":		1000000,	
+		"MaxRecommitInterval":		15000000,	
+		"IntervalAdjustRatio":		0.1,
+		"IntervalAdjustBias":		200000000.0,
+		"StaleThreshold":			7,
+		"DefaultCommitRatio":		0.95,
+		"TxPool":{ 
+			"NoLocals":		 false,
+			"Journal":   	 "transactions.rlp",
+			"Rejournal": 	 3600000000000,
+			"PriceLimit": 	 1,
+			"PriceBump":  	 10,
+			"AccountSlots":	 16,
+			"GlobalSlots":   4096,
+			"AccountQueue":  64,
+			"GlobalQueue":   1024,
+			"GlobalTxCount": 3000,
+			"Lifetime": 	 10800000000000,
+			"ChainHeadChanSize": 250,
+			"TxExtBufferSize":	 4096
+		},
+		"GPO":{
+			"Blocks":       20,
+			"Percentile":   60,
+			"Default":   	100000000,
+			"MaxPrice":		500000000000
+		},
+		"EnablePreimageRecording":false,
+		"DocRoot":				  "",
+		"EWASMInterpreter":		  "",
+		"EVMInterpreter":		  "",
+		"MPCPool":{
+			"MPCEnable":   false,
+			"NoLocals":	   false,
+			"Journal":     "mpc_transactions.rlp",
+			"Rejournal":   4000000000,	
+			"GlobalQueue": 1024,
+			"Lifetime":    10800000000000,
+			"LocalRpcPort":0,
+			"IceConf":	   "",
+			"MpcActor":	   "0x0000000000000000000000000000000000000000"
+		},
+		"VCPool":{
+			"VCEnable":    false,
+			"NoLocals":	   false,
+			"Journal":     "VC_transactions.rlp",
+			"Rejournal":   4000000000,
+			"GlobalQueue": 1024,
+			"Lifetime":    10800000000000,
+			"LocalRpcPort":0,
+			"IceConf":	   "",
+			"VcActor":	   "0x0000000000000000000000000000000000000000",
+			"VcPassword":  ""
+		},
+		"Debug":				 false
+	},
+	"ssh":{
+		"MaxMessageSize":                        1048576,
+		"MinimumAcceptedPOW":                    0.2,
+		"RestrictConnectionBetweenLightClients": true
+	},
+	"node":{
+		"Name":				"platon",
+		"UserIdent":		"",
+		"Version":			"0.7.0-stable",
+		"DataDir":        	"",
+		"KeyStoreDir":		"",
+		"UseLightweightKDF":false,
+		"NoUSB":			false,
+		"IPCPath":			"platon.ipc",
+		"HTTPHost":			"",
+		"HTTPPort":       	6789,
+		"HTTPCors":			[],
+		"HTTPModules":    	["net","web3","eth","ssh"],
+		"HTTPVirtualHosts": ["localhost"],
+		"HTTPTimeouts": {
+			"ReadTimeout": 	30000000000,
+			"WriteTimeout": 30000000000,
+			"IdleTimeout":  120000000000
+		},
+		"WSHost":			"",
+		"WSPort":           6790,
+		"WSOrigins":		[],
+		"WSExposeAll":		false,
+		"WSModules":        ["net","web3","eth","ssh"],
+		"P2P":{
+			"MaxPeers":   	50,
+			"MaxConsensusPeers": 75,
+			"MaxPendingPeers":0,
+			"DialRatio": 0,
+			"NoDiscovery": false,
+			"DiscoveryV5": false,
+			"Name":		   "",
+			"BootstrapNodes":[],
+			"BootstrapNodesV5":[],
+			"StaticNodes":[],
+			"TrustedNodes":[],
+			"NodeDatabase": "",
+			"ListenAddr": 	":16789",
+			"NoDial": false,
+			"EnableMsgEvents": false
+		}
+	},
+	"Ethstats":{
+		"URL":""
+	},
+	"Dashboard":{
+		"Host":    "localhost",
+		"Port":    8080,
+		"Refresh": 5000000000
+	},
+	"EconomicModel":{
+		"Staking":{
+			"StakeThreshold":				1000000000000000000000000,
+			"DelegateThreshold":			10,
+			"ConsValidatorNum":				25,
+			"EpochValidatorNum":			101,
+			"ShiftValidatorNum":			8,
+			"EpochSize":					88,
+			"HesitateRatio":				1,
+			"EffectiveRatio":				1,
+			"ElectionDistance":				20,
+			"ConsensusSize":				250,
+			"UnStakeFreezeRatio":			1,
+			"PassiveUnDelegateFreezeRatio":	0,
+			"ActiveUnDelegateFreezeRatio":	0
+		},
+		"Slashing":{
+			"BlockAmountLow":			8,
+			"BlockAmountHigh":			5,
+			"BlockAmountLowSlashing":	10,
+			"BlockAmountHighSlashing":	20,
+			"DuplicateSignNum":			2,
+			"DuplicateSignLowSlashing":	10,
+			"DuplicateSignHighSlashing":10
+		},
+		"EpochsPerYear":1
+	}
+}
