+++ utils/platon_lib/ppos.py
@@ -23,7 +23,7 @@ def get_sleep_time(number):
         return total_time - number + i
 
 class Ppos:
-    def __init__(self, url, address, chainid=101,
+    def __init__(self, url, address, chainid,
                  privatekey=conf.PRIVATE_KEY):
         self.web3 = connect_web3(url)
         if not self.web3.isConnected():
@@ -35,22 +35,26 @@ class Ppos:
         self.gas = "0x6fffffff"
         self.chainid = chainid
 
-    def get_result(self, tx_hash):
+
+    def get_result(self, tx_hash,func_name):
         result = self.eth.waitForTransactionReceipt(tx_hash)
         print(result)
         """查看eventData"""
         data = result['logs'][0]['data']
         if data[:2] == '0x':
             data = data[2:]
-        # print(data)
+        print(data)
         data_bytes = rlp.decode(bytes.fromhex(data))[0]
         event_data = bytes.decode(data_bytes)
         event_data = json.loads(event_data)
-        # print(event_data)
-        return event_data
+        print(event_data)
+        return func_name,event_data
 
-    def send_raw_transaction(self, data, from_address, to_address, gasPrice, gas,value):
+    def send_raw_transaction(self, data, from_address, to_address, gasPrice, gas,value,privatekey=None):
         nonce = self.eth.getTransactionCount(from_address)
+        print('nonce:',nonce)
+        if not privatekey:
+            privatekey = self.privatekey
         if value > 0:
             transaction_dict = {
                 "to": to_address,
@@ -59,8 +63,9 @@ class Ppos:
                 "nonce": nonce,
                 "data": data,
                 "chainId": self.chainid,
-                "value": self.web3.toWei(value, "ether"),
+                "value": self.web3.toWei(value, "ether")
             }
+
         else:
             transaction_dict = {
                 "to": to_address,
@@ -71,14 +76,14 @@ class Ppos:
                 "chainId": self.chainid
             }
         signedTransactionDict = self.eth.account.signTransaction(
-            transaction_dict, self.privatekey
+            transaction_dict, privatekey
         )
         data = signedTransactionDict.rawTransaction
         result = HexBytes(self.eth.sendRawTransaction(data)).hex()
         return result
 
-    def createStaking(self, typ, benifitAddress, nodeId,externalId, nodeName, website, details,
-                      amount,programVersion,from_address=None, gasPrice=None, gas=None):
+    def createStaking(self, typ, benifitAddress, nodeId,externalId, nodeName, website, details, amount,value,programVersion,privatekey,
+                from_address=None, gasPrice=None, gas=None):
         '''
         createStaking ：发起质押
         :param typ:  uint16(2bytes)
@@ -98,6 +103,7 @@ class Ppos:
             data:[]
             ErrMsg: string
         '''
+        to_address = "0x1000000000000000000000000000000000000002"
         if benifitAddress[:2] == '0x':
             benifitAddress = benifitAddress[2:]
         data = HexBytes(rlp.encode([rlp.encode(int(1000)),
@@ -112,13 +118,13 @@ class Ppos:
         if not gasPrice :
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000002"
-        result = self.send_raw_transaction(data, from_address, to_address, gasPrice,gas,0)
-        # print(result)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice,gas,value,privatekey)
+        print(result)
+        return self.get_result(result, self.createStaking.__name__)
 
-    def updateStakingInfo(self, benifitAddress, nodeId,externalId, nodeName, website, details,
+    def updateStakingInfo(self, benifitAddress, nodeId,externalId, nodeName, website, details,privatekey,
                 from_address=None, gasPrice=None , gas=None):
         '''
         Description : 修改质押信息
@@ -137,6 +143,7 @@ class Ppos:
             ErrMsg: string
 
         '''
+        to_address = "0x1000000000000000000000000000000000000002"
         if benifitAddress[:2] == '0x':
             benifitAddress = benifitAddress[2:]
         data = HexBytes(rlp.encode([rlp.encode(int(1001)),
@@ -149,12 +156,12 @@ class Ppos:
         if not gasPrice:
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000002"
-        result = self.send_raw_transaction(data, from_address, to_address, gasPrice,gas,0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice,gas,0,privatekey)
+        return self.get_result(result, self.updateStakingInfo.__name__)
 
-    def addStaking(self,nodeId,typ,amount,from_address=None, gasPrice=None, gas=None):
+    def addStaking(self,nodeId,typ,amount,privatekey,from_address=None, gasPrice=None, gas=None):
         '''
         Description : 增持质押
         :param nodeId: 64bytes
@@ -168,6 +175,7 @@ class Ppos:
             data:[]
             ErrMsg: string
         '''
+        to_address = "0x1000000000000000000000000000000000000002"
         data = HexBytes(rlp.encode([rlp.encode(int(1002)),
                                     rlp.encode(bytes.fromhex(nodeId)),
                                     rlp.encode(int(typ)),
@@ -177,12 +185,12 @@ class Ppos:
         if not gasPrice:
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000002"
-        result = self.send_raw_transaction(data, from_address, to_address,gasPrice, gas,0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address,gasPrice, gas,0,privatekey)
+        return self.get_result(result, self.addStaking.__name__)
 
-    def unStaking(self,nodeId,from_address=None, gasPrice=None, gas=None):
+    def unStaking(self,nodeId,privatekey,from_address=None, gasPrice=None, gas=None):
         '''
         Description : 撤销质押
         :param nodeId: 64bytes
@@ -194,18 +202,19 @@ class Ppos:
             data:[]
             ErrMsg: string
         '''
+        to_address = "0x1000000000000000000000000000000000000002"
         data = rlp.encode([rlp.encode(int(1003)),rlp.encode(bytes.fromhex(nodeId))])
         if not from_address:
             from_address = self.address
         if not gasPrice:
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000002"
-        result = self.send_raw_transaction(data, from_address, to_address,gasPrice, gas,0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address,gasPrice, gas,0,privatekey)
+        return self.get_result(result, self.unStaking.__name__)
 
-    def delegate(self,typ,nodeId,amount,from_address=None, gasPrice=None , gas=None):
+    def delegate(self,typ,nodeId,amount,privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         Description :发起委托
         :param typ: uint16(2bytes)
@@ -219,6 +228,7 @@ class Ppos:
             data:[]
             ErrMsg: string
         '''
+        to_address = "0x1000000000000000000000000000000000000002"
         data = rlp.encode([rlp.encode(int(1004)),
                            rlp.encode(int(typ)),
                            rlp.encode(bytes.fromhex(nodeId)),
@@ -228,12 +238,12 @@ class Ppos:
         if not gasPrice:
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000002"
-        result = self.send_raw_transaction(data, from_address, to_address,gasPrice, gas,0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address,gasPrice, gas,0,privatekey)
+        return self.get_result(result, self.delegate.__name__)
 
-    def unDelegate(self,stakingBlockNum,nodeId,amount,from_address=None, gasPrice=None , gas=None):
+    def unDelegate(self,stakingBlockNum,nodeId,amount,privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         Description :减持/撤销委托(全部减持就是撤销)
         :param stakingBlockNum: uint64(8bytes)
@@ -244,6 +254,7 @@ class Ppos:
         :param gas:
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000002"
         data = rlp.encode([rlp.encode(int(1005)),rlp.encode(int(stakingBlockNum)),
                            rlp.encode(bytes.fromhex(nodeId)),rlp.encode(self.web3.toWei(amount, 'ether'))])
         if not from_address:
@@ -251,10 +262,10 @@ class Ppos:
         if not gasPrice:
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000002"
-        result = self.send_raw_transaction(data, from_address, to_address,gasPrice, gas,0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address,gasPrice, gas,0,privatekey)
+        return self.get_result(result, self.unDelegate.__name__)
 
     def getVerifierList(self):
         '''
@@ -319,7 +330,6 @@ class Ppos:
         recive = str(recive, encoding="utf8")
         recive = recive.replace('\\','').replace('"[','[').replace(']"',']')
         recive = json.loads(recive)
-        print(recive)
         return recive
 
     def getCandidateList(self):
@@ -358,7 +368,6 @@ class Ppos:
         recive = str(recive, encoding="utf8")
         recive = recive.replace('\\','').replace('"[','[').replace(']"',']')
         recive = json.loads(recive)
-        print(recive)
         return recive
 
     def getDelegateListByAddr(self,addr):
@@ -457,7 +466,7 @@ class Ppos:
         return recive
 
 #################################治理###############################
-    def submitText(self,verifier,githubID,topic,desc,url,endVotingBlock,from_address=None, gasPrice=None , gas=None):
+    def submitText(self,verifier,githubID,topic,desc,url,endVotingBlock,privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         提交文本提案
         :param verifier: 64bytes
@@ -470,6 +479,7 @@ class Ppos:
         :param gas:
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000005"
         data = rlp.encode([rlp.encode(int(2000)),rlp.encode(bytes.fromhex(verifier)),
                                     rlp.encode(githubID),rlp.encode(topic),
                                     rlp.encode(desc), rlp.encode(url),rlp.encode(endVotingBlock)])
@@ -478,12 +488,12 @@ class Ppos:
         if not gasPrice :
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000005"
-        result = self.send_raw_transaction(data, from_address, to_address, gasPrice,gas,0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice,gas,0,privatekey)
+        return self.get_result(result, self.submitText.__name__)
 
-    def submitVersion(self,verifier,githubID,topic,desc,url,newVersion,endVotingBlock,activeBlock,
+    def submitVersion(self,verifier,githubID,topic,desc,url,newVersion,endVotingBlock,activeBlock,privatekey,
                       from_address=None, gasPrice=None , gas=None):
         '''
         提交升级提案
@@ -497,6 +507,7 @@ class Ppos:
         :param activeBlock: uint64
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000005"
         data = rlp.encode([rlp.encode(int(2001)), rlp.encode(bytes.fromhex(verifier)),
                            rlp.encode(githubID), rlp.encode(topic),rlp.encode(desc),
                            rlp.encode(url),rlp.encode(int(newVersion)),rlp.encode(int(endVotingBlock)),
@@ -506,13 +517,13 @@ class Ppos:
         if not gasPrice :
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000005"
-        result = self.send_raw_transaction(data, from_address, to_address, gasPrice,gas,0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice,gas,0,privatekey)
+        return self.get_result(result, self.submitVersion.__name__)
 
     def submitParam(self,verifier,githubID,topic,desc,url,endVotingBlock,paramName,currentValue,newValue,
-                    from_address=None, gasPrice=None , gas=None):
+                    privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         提交参数提案
         :param verifier: 64bytes
@@ -527,6 +538,7 @@ class Ppos:
         :param gas:
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000005"
         data = rlp.encode([rlp.encode(int(2002)),rlp.encode(bytes.fromhex(verifier)),
                            rlp.encode(githubID),rlp.encode(topic),
                            rlp.encode(desc),rlp.encode(url),rlp.encode(int(endVotingBlock)),
@@ -536,12 +548,12 @@ class Ppos:
         if not gasPrice :
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000005"
-        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0, privatekey)
+        return self.get_result(result, self.submitParam.__name__)
 
-    def vote(self,verifier,proposalID,option,programVersion,from_address=None, gasPrice=None , gas=None):
+    def vote(self,verifier,proposalID,option,programVersion,privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         给提案投票
         :param verifier: 64bytes
@@ -552,6 +564,8 @@ class Ppos:
         :param gas:
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000005"
+
         data = rlp.encode([rlp.encode(int(2003)),rlp.encode(bytes.fromhex(verifier)),
                            rlp.encode(int(programVersion)),
                            rlp.encode(proposalID),rlp.encode(option)])
@@ -560,12 +574,12 @@ class Ppos:
         if not gasPrice :
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000005"
-        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0, privatekey)
+        return self.get_result(result, self.vote.__name__)
 
-    def declareVersion(self,activeNode,version,from_address=None, gasPrice=None , gas=None):
+    def declareVersion(self,activeNode,version,privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         版本声明
         :param activeNode: 64bytes
@@ -575,6 +589,8 @@ class Ppos:
         :param gas:
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000005"
+
         data = rlp.encode([rlp.encode(int(2004)), rlp.encode(bytes.fromhex(activeNode)),
                            rlp.encode(int(version))])
         if not from_address:
@@ -582,30 +598,32 @@ class Ppos:
         if not gasPrice :
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000005"
-        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0,privatekey)
+        return self.get_result(result, self.declareVersion.__name__)
 
-    def getProposal(self,proposalID,from_address=None, gasPrice=None , gas=None):
+    def getProposal(self,proposalID,privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         查询提案
         :param proposalID: common.Hash
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000005"
+
         data = rlp.encode([rlp.encode(int(2100)),rlp.encode(proposalID)])
         if not from_address:
             from_address = self.address
         if not gasPrice :
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000005"
-        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0,privatekey)
         # print(result)
-        return self.get_result(result)
+        return self.get_result(result, self.getProposal.__name__)
 
-    def getTallyResult(self,proposalID,from_address=None, gasPrice=None , gas=None):
+    def getTallyResult(self,proposalID,privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         查询提案结果
         :param proposalID: common.Hash
@@ -614,68 +632,43 @@ class Ppos:
         :param gas:
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000005"
+
         data = rlp.encode([rlp.encode(int(2101)),rlp.encode(str(proposalID))])
         if not from_address:
             from_address = self.address
         if not gasPrice :
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000005"
-        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0,privatekey)
         print(result)
-        return self.get_result(result)
+        return self.get_result(result, self.getTallyResult.__name__)
 
-    def listProposal(self):
+    def listProposal(self,privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         查询提案列表
+        :param from_address:
+        :param gasPrice:
+        :param gas:
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000005"
+
         data = rlp.encode([rlp.encode(int(2102))])
-        to_address = "0x1000000000000000000000000000000000000005"
-        recive = self.eth.call({
-            "from": self.address,
-            "to": to_address,
-            "data": data
-        })
-        recive = str(recive, encoding="utf8")
-        recive =json.loads(recive)
-        return recive
-
-    def getActiveVersion(self):
-        """
-        查询节点的链生效版本
-        """
-        data = rlp.encode([rlp.encode(int(2103))])
-        to_address = "0x1000000000000000000000000000000000000005"
-        recive = self.eth.call({
-            "from": self.address,
-            "to": to_address,
-            "data": data
-        })
-        recive = str(recive, encoding="utf8")
-        recive =json.loads(recive)
-        # print(recive)
-        return recive
-
-    def listParam(self):
-        """
-        查询可治理参数列表
-        """
-        data = rlp.encode([rlp.encode(int(2105))])
-        to_address = "0x1000000000000000000000000000000000000005"
-        recive = self.eth.call({
-            "from": self.address,
-            "to": to_address,
-            "data": data
-        })
-        recive = str(recive, encoding="utf8")
-        recive =json.loads(recive)
-        # print(recive)
-        return recive
+        if not from_address:
+            from_address = self.address
+        if not gasPrice :
+            gasPrice = self.gasPrice
+        if not gas:
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0,privatekey)
+        return self.get_result(result, self.listProposal.__name__)
 
 ############################举报惩罚###############################################################
-    def ReportMutiSign(self,data,from_address=None, gasPrice=None , gas=None):
+    def ReportMutiSign(self,data,privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         举报双签
         :param data: string
@@ -684,18 +677,20 @@ class Ppos:
         :param gas:
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000004"
+
         data_ = rlp.encode([rlp.encode(data)])
         if not from_address:
             from_address = self.address
         if not gasPrice:
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000004"
-        result = self.send_raw_transaction(data_, from_address, to_address, gasPrice, gas, 0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data_, from_address, to_address, gasPrice, gas, 0,privatekey)
+        return self.get_result(result, self.ReportMutiSign.__name__)
 
-    def CheckMutiSign(self,typ,addr,blockNumber,from_address=None, gasPrice=None , gas=None):
+    def CheckMutiSign(self,typ,addr,blockNumber,privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         查询节点是否已被举报过多签
         :param typ: uint8(1byte)
@@ -706,6 +701,8 @@ class Ppos:
         :param gas:
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000004"
+
         if addr[:2] == '0x':
             addr = addr[2:]
         data = rlp.encode([rlp.encode(int(3001)),rlp.encode(int(typ)),
@@ -715,13 +712,13 @@ class Ppos:
         if not gasPrice:
             gasPrice = self.gasPrice
         if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000004"
-        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0)
-        return self.get_result(result)
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0,privatekey)
+        return self.get_result(result, self.CheckMutiSign.__name__)
 #######################################锁仓###############################################
 
-    def CreateRestrictingPlan(self,account,plan,from_address=None, gasPrice=None , gas=None):
+    def CreateRestrictingPlan(self,account,plan,privatekey,from_address=None, gasPrice=None , gas=None):
         '''
         创建锁仓计划
         :param account: 20bytes
@@ -731,27 +728,28 @@ class Ppos:
         :param gas:
         :return:
         '''
+        to_address = "0x1000000000000000000000000000000000000001"
         if account[:2] == '0x':
             account = account[2:]
         plan_list = []
         for dict_ in plan:
-            v = [rlp.encode(dict_[k]) for k in dict_]
+            v = [dict_[k] for k in dict_]
             plan_list.append(v)
+        rlp_list =rlp.encode(plan_list)
         data = rlp.encode([rlp.encode(int(4000)),
                            rlp.encode(bytes.fromhex(account)),
-                           rlp.encode(plan_list)])
+                           rlp_list])
         if not from_address:
             from_address = self.address
         if not gasPrice:
             gasPrice = self.gasPrice
-        if not gas:
-            gas = self.gas
-        to_address = "0x1000000000000000000000000000000000000001"
-        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0)
-        print(result)
-        return self.get_result(result)
+        if not gas :
+            transactiondict = {"to": to_address, "data": data}
+            gas = self.eth.estimateGas (transactiondict)
+        result = self.send_raw_transaction(data, from_address, to_address, gasPrice, gas, 0,privatekey)
+        return self.get_result(result, self.CreateRestrictingPlan.__name__)
 
-    def GetRestrictingInfo(self,account):
+    def GetRestrictingInfo(self,account ):
         '''
         获取锁仓信息
         :param account: 20bytes
@@ -771,82 +769,16 @@ class Ppos:
         })
         recive = str(recive, encoding="ISO-8859-1")
         recive = json.loads(recive)
-        print(recive)
+        # recive = recive[10:]
+        # recive = eval(recive)
+        # print(recive)
         return recive
+
 
 
 if __name__ == '__main__':
     address = '0x493301712671Ada506ba6Ca7891F436D29185821'
-    # address = "0xb37F459F9F0632C57f1247d495a1D96a3d5b3c15"
     # p = Ppos( 'http://10.10.8.157:6789',address ,'88888888')
-    p = Ppos('http://192.168.9.221:6789', address,101)
-    # Ppos('http://10.10.8.157:6789', address, 102)
-    typ= 0
-    benifitAddress = '0x493301712671Ada506ba6Ca7891F436D29185821'
-    nodeId = 'a5d6f3ac90e843e74cc9e1477b32776ae223351d5cb2654397a653c635bc3e7de73fe6a6f77c20af4a693e9e244df6764a40d396930431527a16c989f129ad89'
-    externalId = 'sfsf'
-    nodeName ='ffswfv'
-    website='ffaf'
-    details = 'effs'
-    amount= 10000000
-    programVersion= 1792
-    stakingBlockNum = 66
-    privatekey = '0xa11859ce23effc663a9460e332ca09bd812acc390497f8dc7542b6938e13f8d7'
-    # p.addStaking(nodeId,typ,amount)
-    # p.GetRestrictingInfo(benifitAddress)
-    # p.updateStakingInfo(benifitAddress, nodeId,externalId, nodeName, website, details)
-    # p.createStaking(typ, address, nodeId,externalId, nodeName, website, details, amount,programVersion)
-    # p.getVerifierList()
-    # p.getValidatorList()
-    # p.getCandidateList()
-    # p.getDelegateListByAddr(address)
-    # p.getDelegateInfo(stakingBlockNum,benifitAddress,nodeId)
-    # p.getCandidateInfo(nodeId)
-    # p.GetRestrictingInfo(address)
-    # proposalID = 66666666666666666666666666666666
-    # p.getTallyResult(proposalID)
-    # web3 = Web3(Web3.HTTPProvider('http://10.10.8.157:6789'))
-    # v1 = web3.toWei(80, 'ether')
-    # v2 = web3.toWei(66, 'ether')
-    # p.addStaking(nodeId,typ,amount)
-    # p.getVerifierList()
-    # p.getDelegateInfo()
-    # p.GetRestrictingInfo(address)
-    verifier = 'f71e1bc638456363a66c4769284290ef3ccff03aba4a22fb60ffaed60b77f614bfd173532c3575abe254c366df6f4d6248b929cb9398aaac00cbcc959f7b2b7c'
-    # verifier = "a28b52294324f17a8e5e15da2c1562494303d694a9ac6ca02c2ae78fd93af69bb454a711883c23d9a96e38b88e389fbb6225fc6578cb22b4905520c8fbd000c3"
-    githubID='66'
-    topic = '888gergregreggregergregweferfwregweagf'
-    desc = 'helloeferfgwrgrgwrgregregregregw'
-    url = 'baidu.com'
-    newVersion = 1354354
-    endVotingBlock = 55
-    activeBlock = 66
-    currentValue= '66'
-    newValue = '11'
-    activeNode = verifier
-    version = 88
-    option = 'y'
-    # p.getDelegateListByAddr()
-    # p.submitText(verifier,githubID,topic,desc,url,endVotingBlock)
-    proposalID = 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
-    paramName ='wuyiqin'
-    # p.getProposal(proposalID)
-    # p.submitText(verifier, githubID, topic, desc, url, endVotingBlock)
-    # p.submitVersion(verifier,githubID,topic,desc,url,newVersion,endVotingBlock,activeBlock)
-    # p.submitParam(verifier,githubID,topic,desc,url,endVotingBlock,paramName,currentValue,newValue)
-    # p.vote(verifier,proposalID,option,10)
-    # p.declareVersion(activeNode,version)
-    # p.getProposal(proposalID)
-    # p.getTallyResult(proposalID)
-    # p.listProposal()
-    # p.submitText(verifier, githubID, topic, desc, url, endVotingBlock)
-    plan =  [{"epoch": 12, "amount": 45}, {"epoch": 24, "amount": 90}]
-    p.listParam()
-    # p.CreateRestrictingPlan(benifitAddress,plan,"0xa11859ce23effc663a9460e332ca09bd812acc390497f8dc7542b6938e13f8d7")
-
-
-
-
-
-
-
+    p = Ppos('http://10.10.8.157:6789', address,101)
+    p.get_result('0x03ecc55232cc5e65f0f95390fa7bb5cd8f3d74439bc607cd36cdac37a861d882','CreateRestrictingPlan')
+    p.GetRestrictingInfo()
